All tasks done

+++ app/Http/Controllers/CurrencyController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Currency;
 use Illuminate\Http\Request;
 use App\Http\Requests\CurrencyRequest;
+use Illuminate\Support\Facades\Gate;
 
 class CurrencyController extends Controller
 {
@@ -21,7 +22,7 @@ class CurrencyController extends Controller
     public function index()
     {
         $currencies = Currency::all();
-        return view('currencies.index', ['currencies' => $currencies->toArray()]);
+        return view('currencies.index', ['currencies' => $currencies]);
     }
 
     /**
@@ -31,6 +32,9 @@ class CurrencyController extends Controller
      */
     public function create()
     {
+        if (Gate::denies('create', Currency::class)) {
+            return redirect('/');
+        }
         return view('currencies.add');
     }
 
@@ -42,11 +46,14 @@ class CurrencyController extends Controller
      */
     public function store(CurrencyRequest $request)
     {
+        if (Gate::denies('create', Currency::class)) {
+            return redirect('/');
+        }
         $data = $request->only(['title', 'short_name', 'logo_url', 'price']);
         $currency = new Currency($data);
         $currency->save();
         $currencies = Currency::all();
-        return view('currencies.index', ['currencies' => $currencies->toArray()]);
+        return view('currencies.index', ['currencies' => $currencies]);
     }
 
     /**
@@ -58,7 +65,10 @@ class CurrencyController extends Controller
     public function show($id)
     {
         $currency = Currency::find($id);
-        return view('currencies.show', ['currency' => $currency->toArray()]);
+        if (Gate::denies('view', $currency)) {
+            return redirect('/');
+        }
+        return view('currencies.show', ['currency' => $currency]);
     }
 
     /**
@@ -70,7 +80,10 @@ class CurrencyController extends Controller
     public function edit($id)
     {
         $currency = Currency::find($id);
-        return view('currencies.edit', ['currency' => $currency->toArray()]);
+        if ($currency === null || Gate::denies('update', $currency)) {
+            return redirect('/');
+        }
+        return view('currencies.edit', ['currency' => $currency]);
     }
 
     /**
@@ -83,6 +96,9 @@ class CurrencyController extends Controller
     public function update(CurrencyRequest $request, $id)
     {
         $currency = Currency::find($id);
+        if ($currency === null || Gate::denies('update', $currency)) {
+            return redirect('/');
+        }
         $currency->title = $request->input('title');
         $currency->short_name = $request->input('short_name');
         $currency->logo_url = $request->input('logo_url');
@@ -100,6 +116,9 @@ class CurrencyController extends Controller
     public function destroy($id)
     {
         $currency = Currency::find($id);
+        if ($currency === null || Gate::denies('delete', $currency)) {
+            return redirect('/');
+        }
         $currency->delete();
         return redirect()->route('currencies.index');
     }
